Bug
Added a new command telling users to reach out to either Kgreene829(me) or Tuscansota(you) through discord DMs if they find a bug in the bot. I think the code should be bug free but in converting everything to take team ID and adding in the names of teams I may have missed something or a certain team may not be working. Now if someone finds a bug with a certain team not working we will be notified of it and I can fix it quickly. I also said to message me first since the bugs are probably going to be something with the team logo not showing up or a certain team not working which is just a small fix that I can do quickly